mmmmmm idunno why im getting these errors

+++ hydroponic_automation.ino
@@ -1,7 +1,7 @@
 #include <LiquidCrystal.h>
-#include "SensorAutomation.h"
+#include <SensorAutomation.h>
 
-const double REG_EC_INTERVAL = 60*60*1000; // Regular interval for EC loop is 1 hour
+const unsigned long int REG_EC_INTERVAL = 60*60*1000; // Regular interval for EC loop is 1 hour
 const double REG_PH_INTERVAL = 10*1000; // Regular pH loop will occur 10 minutes after EC loop
 
 int ec_read = A0;
